requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-retry==2.9.1
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.45.0
anyio==4.8.0
attrs==25.1.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.9.0
boto3==1.36.5
botocore==1.36.5
browser-use==0.1.27
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cohere==5.13.12
colorama==0.4.6
contourpy==1.3.0
courlan==1.3.2
cycler==0.12.1
dateparser==1.2.0
defusedxml==0.7.1
Deprecated==1.2.15
diskcache==5.6.3
distro==1.9.0
elevenlabs==1.55.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.8
fastavro==1.10.0
filelock==3.17.0
filetype==1.2.0
fireworks-ai==0.15.12
Flask==3.1.3
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
geopandas==1.0.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
html2text==2024.2.26
htmldate==1.9.3
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
httpx-ws==0.7.1
huggingface-hub==0.29.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jusText==3.0.1
kaleido==0.2.1
kiwisolver==1.4.7
langchain==0.3.14
langchain-anthropic==0.3.3
langchain-aws==0.2.11
langchain-core==0.3.31
langchain-fireworks==0.2.6
langchain-google-genai==2.0.8
langchain-ollama==0.2.2
langchain-openai==0.3.1
langchain-text-splitters==0.3.5
langsmith==0.2.11
lida==0.0.14
llmx==0.0.21a0
lmnr==0.4.55
lxml==5.3.0
lxml_html_clean==0.4.1
MainContentExtractor==0.0.4
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-venn==1.1.1
mdurl==0.1.2
mizani==0.11.4
monotonic==1.6
multidict==6.1.0
narwhals==1.27.1
networkx==3.2.1
numpy==2.0.2
ollama==0.4.7
openai==1.64.0
openpyxl==3.1.5
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-langchain==0.36.1
opentelemetry-instrumentation-requests==0.50b0
opentelemetry-instrumentation-sqlalchemy==0.50b0
opentelemetry-instrumentation-threading==0.50b0
opentelemetry-instrumentation-urllib3==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-semantic-conventions-ai==0.4.2
opentelemetry-util-http==0.50b0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==11.1.0
playwright==1.49.1
plotly==6.0.0
plotnine==0.13.6
posthog==3.9.3
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
pyee==12.0.0
Pygments==2.19.1
PyJWT==2.12.1
pyogrio==0.10.0
pyparsing==3.2.1
PyPDF2==3.0.1
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
s3transfer==0.11.2
scikit-learn==1.7.2
scipy==1.13.1
seaborn==0.13.2
setuptools==75.8.0
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
starlette==0.45.3
statsmodels==0.14.4
streamlit==1.42.2
streamlit-option-menu==0.4.0
tenacity==9.0.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tld==0.13
tokenizers==0.21.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
trafilatura==2.0.0
twilio==9.10.3
typer==0.15.1
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
ucimlrepo==0.0.7
uritemplate==4.1.1
urllib3==2.3.0
uv==0.5.24
uvicorn==0.34.0
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.6
whisper==1.1.10
wordcloud==1.9.4
wrapt==1.17.2
wsproto==1.2.0
xlrd==2.0.1
yarl==1.18.3
zipp==3.21.0
